Fixed price. Changed from int to float.

+++ src/main/java/com/website/DatabaseInterface.java
@@ -21,7 +21,7 @@ public interface DatabaseInterface extends Remote {
      * @param ID
      * @throws RemoteException
      */
-    List<product> dbFetch(String name, int price_min, int price_max, String type, int ID) throws RemoteException;
+    List<product> dbFetch(String name, float price_min, float price_max, String type, int ID) throws RemoteException;
 
     /**
      * Method to manage database operations such as creating, updating, or deleting entries.
